Extract babel-loader into rules/javascript
- Pass arguments as object
- babel provides support for the latest javascript. To me that is more of an
  implementation detail. From a higher-level, we use babel for javascript. It
  seems like a more logical and meaningful name.
