+++ webpack/webpack.config.js
@@ -14,6 +14,7 @@ const webpack = require('webpack');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const PATHS = require('./paths');
 const image = require('./rules/image');
+const javascript = require('./rules/javascript');
 
 module.exports = (env = '') => {
   const isProd = process.env.NODE_ENV === 'production';
@@ -23,27 +24,6 @@ module.exports = (env = '') => {
   const externals = fs.readdirSync('node_modules')
     .filter(x => ['.bin'].indexOf(x) === -1)
     .reduce((acc, cur) => Object.assign(acc, { [cur]: 'commonjs ' + cur }), {});
-
-  const babelLoader = (production = false, browser = false) => {
-    const obj = {
-      test: /\.js$|\.jsx$/,
-      loader: 'babel-loader',
-      query: {
-        presets: ['es2015', 'react', 'stage-0']
-      },
-      exclude: PATHS.modules
-    };
-    if (production) {
-      obj.query.plugins = [
-        'transform-react-remove-prop-types',
-        'transform-react-constant-elements',
-        'transform-react-inline-elements'
-      ];
-    } else if (browser) {
-      obj.query.presets.unshift('react-hmre');
-    }
-    return obj;
-  };
 
   const cssLoader = (production = false, browser = false) => {
     /*
@@ -141,7 +121,17 @@ module.exports = (env = '') => {
       publicPath: PATHS.public,
       libraryTarget: 'commonjs2'
     },
-    module: { rules: [babelLoader(isProd, false), cssLoader(isProd, false), image()] },
+    module: {
+      rules:
+        [
+          javascript({
+            production: isProd,
+            browser: false
+          }),
+          cssLoader(isProd, false),
+          image()
+      ]
+    },
     resolve,
     plugins: webpackPlugins(true, false)
   };
@@ -156,7 +146,17 @@ module.exports = (env = '') => {
       chunkFilename: '[name].[chunkhash:6].js', // for code splitting. will work without but useful to set
       publicPath: PATHS.public
     },
-    module: { rules: [babelLoader(isProd, true), cssLoader(isProd, true), image()] },
+    module: {
+      rules:
+        [
+          javascript({
+            production: isProd,
+            browser: true
+          }),
+          cssLoader(isProd, true),
+          image()
+        ]
+    },
     resolve,
     plugins: webpackPlugins(true, true)
   };
@@ -175,7 +175,17 @@ module.exports = (env = '') => {
       filename: '[name].js',
       publicPath: PATHS.public
     },
-    module: { rules: [babelLoader(isProd, true), cssLoader(isProd, true), image()] },
+    module: {
+      rules:
+        [
+          javascript({
+            production: isProd,
+            browser: true
+          }),
+          cssLoader(isProd, true),
+          image()
+        ]
+    },
     resolve,
     plugins: webpackPlugins(false, true)
   };
@@ -193,7 +203,17 @@ module.exports = (env = '') => {
       publicPath: PATHS.public,
       libraryTarget: 'commonjs2',
     },
-    module: { rules: [babelLoader(isProd, false), cssLoader(isProd, false), image()] },
+    module: {
+      rules:
+        [
+          javascript({
+            production: isProd,
+            browser: false
+          }),
+          cssLoader(isProd, false),
+          image()
+        ]
+    },
     resolve,
     plugins: webpackPlugins(false, false)
   };
